// Rate-Runner-API/src/modules/profiles/models/CarrierProfile.ts
export class CarrierProfile {
    cr_id: string
    cr_code: string
    carrier_city_name: string
    carrier_country_name: string
    cr_city_id: string
    cr_country_id: string
    cr_name: string
    cr_address: string
    cr_mobile: string
    cr_phone: string
    cr_mail: string
    cr_official_social_id: string
    cr_website: string
    cr_remark: string
    cr_disable: number
}