// Rate-Runner-API/src/modules/profiles/models/CustomerProfile.ts
export class CustomerProfile {
    cus_id: string
    cus_city_id: string
    customer_city_name: string
    customer_country_name: string
    cus_country_id: string
    cus_name: string
    cus_address: string
    cus_type: string
    cus_industry: string
    cus_mobile: string
    cus_phone: string
    cus_mail: string
    cus_official_social_id: string
    cus_website: string
    cus_disable: number
    sale_owners: any[]
}